+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- feat: Add angle unit function `deg()` — and converters from `turn`, `rad`
+  and `grad`
+- feat: Add `VariablePrinter.type`
 
 ## 0.1.0